#2 - [move - user] Request 위치 변경
