+++ src/main/java/me/kimgunwoo/auctionseats/domain/user/dto/request/UserInfoUpdateRequest.java
@@ -1,4 +1,4 @@
-package me.kimgunwoo.auctionseats.domain.user.dto;
+package me.kimgunwoo.auctionseats.domain.user.dto.request;
 
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.Pattern;
